fix: recipe detail load time — eliminate N+1 queries and stale modal styles
- get_interactions_for_recipe now reads from recipe.nutrients (stored JSONB)
  instead of re-computing via RecipeUtils (N+1 get_ingredient_details! per ingredient)
- recipe_browser_live: replace commented-out RecipeUtils.get_nutrients calls with
  recipe.nutrients / recipe.primary_nutrients_size (fixes undefined variable errors)
- recipe_details_component: restore interactions assign (was commented out, leaving
  a dangling assign(:interactions, interactions) that would crash on first render)
- recipe_detail_timing event handler moved to live_helpers macro so all three
  parent LiveViews handle it (pushEvent routes to parent, not component)
- recipe_modal / ingredient_modal overlays changed from bg-zinc-50/90 (white) to
  bg-slate-950/75 backdrop-blur-sm; panels changed from bg-white to bg-slate-900/800
  to match the app's dark slate aesthetic
